Some changes after discussion with Nick 1. $PROB/$PROBLEM records come from the TARGET_CODE block, if the corresponding TARGET_CODE block is not defined, than a record is generated
2. NM-TRAN record names are printed only if corresponding MDL blocks are
defined.
The presence of TARGET_CODE blocks does not influenece on this.

3. Validation for same block names added: names must match a matrix or a
diag block name.

4. The generation of $OMEGA and $SIGMA sections is revised: processing
of variability subblocks is now similar to the processing of statements
of the VARIABILITY blocks, i.e., the content of diag and matrix blocks
is compared with the lists of ETA and EPS variables and blocks with
variables that appear in the ETA set go to $OMEGA section while blocks
with variables that appear in the EPS set go to $SIGMA section.

It seems to work fine (tested warfarin_PK_BOV.mdl), please check!
!!! same blocks are not printed yet

5. "Before" and "after" attributes are allowed in TARGET_CODE blocks,
but not processed

6. Interface:
- MDL Perspective includes the outline view by default
- Plug-in MDL project removed from the wizard - it is not needed, now
there is only a project without any dependencies