1. Classes for the validation of MDL function signatures are added, the signatures of standard functions are represented in the new format. The quick fix module revised to insert IMPORT blocks to the model object.
2. Conversion of MDL use types to PharmML column types corrected: AMT ->
dose, etc.

(These are the changed done last wekk but I had a conflict after the
revision of the structure of the project which I could not resolve by
merging my version with the head (something to do with git ignore
files), so I am now putting them on top of the current master version).