DDMORE-783: Strip off quotes for default $EST method in MDL->NMTRAN:
If there is no TARGET_CODE block for $ESTIMATION then one is generated
from the appropriate parameters etc. in the MDL file. The generated $EST
then looks like: $EST METHOD="CONDITIONAL  INTER" NOABORT

NMTRAN doesn't like the enclosing quotes, so in printDefaultEstimate()
method of Mdl2Nonmem, strip them off if they are present. Not sure I
like this fix though; since the METHOD seems to be a fixed set of
values, I would prefer to have the METHOD being an enum rather than a
string and then the quotes would not appear anyway, but there is string
manipulation going on which would make this a more involved change.