DDMORE-768: Have to create new Feature for JRE bundled with the MDL IDE.
This is an attempt to work round the problem of not being able to update
the MDL IDE SDK Feature from within the MDL IDE, due to it failing due
to the JRE being in use (by the MDL IDE itself) and thus not replaceable
by the update process (even if the JRE hasn't physically changed; the
contents of a Feature are always replaced by the P2 update).

By creating a separate Feature just containing the JRE to be bundled
with the MDL IDE, we can fix the version of the Feature (i.e. not having
a -SNAPSHOT/.qualifier suffix on the version number that resolves to a
timestamp), and so hopefully this MDL IDE SDK JRE Feature will not
attempt to be updated when doing an update check from within the MDL
IDE.

However I don't know if this will work until I try it... (It shouldn't
be any worse...)

This commit will be benign since nothing will be referencing this new
Feature yet.