Validation of the source data file added:
"With DATA_INPUT_VARIABLES, if header=true in SOURCE then map variable
names to Data Source names. If any do not match then throw an error. If
you have > columns in external file then that's OK, but if
DATA_INPUT_VARIABLES aren't matched by 1 column in external file then
error.
If header is not set in SOURCE then ASSUME that DATA_INPUT_VARIABLES are
defined in the column order they appear in the dataset AND there must be
same number of definitions as there are columns."