DDMORE-1554: Refactor Export as PharmML feature to not break Conversion
Exception that was being thrown by MDL->PharmML Converter:
java.lang.SecurityException: class
"org.eclipse.core.runtime.IAdaptable"'s signer information does not
match signer information of other classes in the same package

This error occurs when a standalone instance of the MDL core is
accessing Eclipse internal classes. Here it was caused by the loading of
the MdlGenerator class by AbstractMdlRuntimeModule.bindIGenerator()
which is called by MdlStandaloneSetupGenerated which is how the external
converters create a standalone instance of the MDL core.

Since the on-the-fly PharmML generation has been ditched, I've removed
the ConvertToPharmMLHandler, ConvertToPharmMLHandler_PE and
AbstractConvertToPharmMLHandler classes, and the problematic
MdlGenerator class, refactoring its remaining code into PharmMLExporter.
Since PharmMLExporter is in the mdl.ui package, there is no chance of it
getting ‘accidentally’ loaded by the MDL core standalone setup unlike
what was happening with the MdlGenerator class.

Finally, removed the generator fragment from GenerateMdl.mwe2 so that
the generated AbstractMdlRuntimeModule.bindIGenerator() method, which
was triggering the error, disappears.